refactor: move (auth) route group out of (features)
