+++ src/app/(auth)/auth.setup.ts
